refactor: remove nested ternaries, extend input and button props, and separate conditional rendering into component
